refactor: Section에 대한 service와 controller를 line 내부에 삽입
section에 대한 기능이 line의 내부에서 동작하므로 하나로 합침₩

-
