--- src/main/java/subway/controller/SectionController.java
@@ -1,29 +1,0 @@
-package subway.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import subway.dto.SectionRequest;
-import subway.factory.SectionFactory;
-import subway.service.SectionService;
-
-@RestController
-@RequestMapping("/lines")
-public class SectionController {
-    private final SectionService sectionService;
-
-    public SectionController(SectionService sectionService) {
-        this.sectionService = sectionService;
-    }
-
-    @PostMapping("/{lineId}/sections")
-    public ResponseEntity createSection(@PathVariable Long lineId, @RequestBody SectionRequest sectionRequest) {
-        sectionService.saveSection(SectionFactory.getSection(sectionRequest, lineId));
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{lineId}/sections")
-    public ResponseEntity deleteSection(@PathVariable Long lineId, @RequestParam Long stationId) {
-        sectionService.deleteSection(lineId, stationId);
-        return ResponseEntity.ok().build();
-    }
-}
