+++ src/main/java/subway/domain/Line.java
@@ -8,23 +8,26 @@ public class Line {
     private final String color;
     private Sections sections;
 
-
-    public Line(Long id, String name, String color) {
-        this.id = id;
-        this.color = color;
-        this.name = name;
-    }
-
-    public Line(Long id, String name, String color, Sections sections) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-        this.sections = sections;
-    }
-
     public Line(String name, String color) {
         this.name = name;
         this.color = color;
+        sections = new Sections();
+    }
+
+    public Line(String name, String color, Sections sections) {
+        this(name,color);
+        this.sections = sections;
+    }
+
+    public Line(Long id, String name, String color) {
+        this(name,color);
+        this.id = id;
+        sections = new Sections();
+    }
+
+    public Line(Long id, String name, String color, Sections sections) {
+        this(name,color,sections);
+        this.id = id;
     }
 
     public Long getId() {
