+++ src/main/java/subway/factory/LineFactory.java
@@ -1,6 +1,7 @@
 package subway.factory;
 
 import subway.domain.Line;
+import subway.domain.Sections;
 import subway.dto.LineRequest;
 
 public class LineFactory {
@@ -10,5 +11,9 @@ public class LineFactory {
 
     public static Line getLine(Long lineId, LineRequest lineRequest) {
         return new Line(lineId, lineRequest.getName(), lineRequest.getColor());
+    }
+
+    public static Line getLine(LineRequest lineRequest, Sections sections) {
+        return new Line(lineRequest.getName(), lineRequest.getColor(), sections);
     }
 }
