+++ src/main/java/subway/service/LineService.java
@@ -1,6 +1,7 @@
 package subway.service;
 
 import subway.domain.Line;
+import subway.domain.Sections;
 import subway.dto.LineRequest;
 import subway.dto.LineResponse;
 import subway.domain.Section;
@@ -18,4 +19,12 @@ public interface LineService {
     Line findOne(Long lineId);
 
     void update(Line line);
+
+    void deleteSection(Long lineId, Long stationId);
+
+    void saveSection(Section section);
+
+    void saveSection(Line line, Section section);
+
+    Sections getSectionsByLineId(Long lineId);
 }
